Removed sqlite references Refactor
