--- SampleODBCApplication/Sample-Test1/pch.h
@@ -1,8 +1,0 @@
-//
-// pch.h
-// Header for standard system include files.
-//
-
-#pragma once
-
-#include "gtest/gtest.h"
