--- SampleODBCApplication/Sample-Test1/test.cpp
@@ -1,6 +1,0 @@
-#include "pch.h"
-
-TEST(TestCaseName, TestName) {
-  EXPECT_EQ(1, 1);
-  EXPECT_TRUE(true);
-}
