+++ odbc-ipfs-sample-application/SampleODBCApplication/SampleODBCApplication.h
@@ -9,9 +9,4 @@
 void listSQLDrivers();
 void listDataSources();
 SQLHDBC conn(SQLHENV* env);
-void getInfo(SQLHDBC dbc);
 void getData(SQLHDBC dbc);
-void getData2(SQLHDBC dbc);
-
-void printSQLCharString(SQLWCHAR* sqlCharString, SQLSMALLINT length);
-SQLWCHAR* stringToSQLWCHAR(char* str);
