+++ odbc-ipfs-sample-application/SampleODBCApplication/SampleODBCApplication.vcxproj
@@ -142,13 +142,9 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="SampleODBCApplication.cpp" />
-    <ClCompile Include="sqlite64\shell.c" />
-    <ClCompile Include="sqlite64\sqlite3.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="SampleODBCApplication.h" />
-    <ClInclude Include="sqlite64\sqlite3.h" />
-    <ClInclude Include="sqlite64\sqlite3ext.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
