+++ odbc-ipfs-sample-application/SampleODBCApplication/SampleODBCApplication.vcxproj.filters
@@ -18,22 +18,10 @@
     <ClCompile Include="SampleODBCApplication.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="sqlite64\shell.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="sqlite64\sqlite3.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="SampleODBCApplication.h">
       <Filter>Source Files</Filter>
-    </ClInclude>
-    <ClInclude Include="sqlite64\sqlite3.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="sqlite64\sqlite3ext.h">
-      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
