+++ odbc-ipfs-sample-application/SampleODBCApplication/sqlite64/sqlite3.lib
